requirements.txt
PyYAML
tqdm
comet-ml
scikit-learn
matplotlib

# PyTorch
# --find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.8.1+cu111 
torchvision==0.9.1+cu111


# pip install torch==1.8.1+cu111 torchvision==0.9.1+cu111 torchaudio==0.8.1 -f https://download.pytorch.org/whl/torch_stable.html